+++ bootique-jdbc-hikaricp-instrumented/src/main/java/io/bootique/jdbc/instrumented/hikaricp/JdbcHikariCPInstrumentedModuleProvider.java
@@ -3,7 +3,7 @@ package io.bootique.jdbc.instrumented.hikaricp;
 import com.google.inject.Module;
 import io.bootique.BQModule;
 import io.bootique.BQModuleProvider;
-import io.bootique.jdbc.hikaricp.HikariCPModuleProvider;
+import io.bootique.jdbc.hikaricp.JdbcHikariCPModuleProvider;
 import io.bootique.metrics.MetricsModuleProvider;
 import io.bootique.metrics.health.HealthCheckModuleProvider;
 
@@ -11,11 +11,11 @@ import java.util.Collection;
 
 import static java.util.Arrays.asList;
 
-public class HikariCPInstrumentedModuleProvider implements BQModuleProvider {
+public class JdbcHikariCPInstrumentedModuleProvider implements BQModuleProvider {
 
     @Override
     public Module module() {
-        return new HikariCPInstrumentedModule();
+        return new JdbcHikariCPInstrumentedModule();
     }
 
     @Override
@@ -28,7 +28,7 @@ public class HikariCPInstrumentedModuleProvider implements BQModuleProvider {
     @Override
     public Collection<BQModuleProvider> dependencies() {
         return asList(
-                new HikariCPModuleProvider(),
+                new JdbcHikariCPModuleProvider(),
                 new MetricsModuleProvider(),
                 new HealthCheckModuleProvider()
         );