+++ bootique-jdbc-tomcat-instrumented/src/main/java/io/bootique/jdbc/instrumented/tomcat/JdbcTomcatInstrumentedModuleProvider.java
@@ -3,7 +3,7 @@ package io.bootique.jdbc.instrumented.tomcat;
 import com.google.inject.Module;
 import io.bootique.BQModule;
 import io.bootique.BQModuleProvider;
-import io.bootique.jdbc.tomcat.TomcatJdbcModuleProvider;
+import io.bootique.jdbc.tomcat.JdbcTomcatModuleProvider;
 import io.bootique.metrics.MetricsModuleProvider;
 import io.bootique.metrics.health.HealthCheckModuleProvider;
 
@@ -11,11 +11,11 @@ import java.util.Collection;
 
 import static java.util.Arrays.asList;
 
-public class TomcatInstrumentedJdbcModuleProvider implements BQModuleProvider {
+public class JdbcTomcatInstrumentedModuleProvider implements BQModuleProvider {
 
     @Override
     public Module module() {
-        return new TomcatInstrumentedJdbcModule();
+        return new JdbcTomcatInstrumentedModule();
     }
 
     @Override
@@ -28,7 +28,7 @@ public class TomcatInstrumentedJdbcModuleProvider implements BQModuleProvider {
     @Override
     public Collection<BQModuleProvider> dependencies() {
         return asList(
-                new TomcatJdbcModuleProvider(),
+                new JdbcTomcatModuleProvider(),
                 new MetricsModuleProvider(),
                 new HealthCheckModuleProvider()
         );