+++ bootique-jdbc-tomcat-instrumented/src/main/java/io/bootique/jdbc/instrumented/tomcat/TomcatMetricsInitializer.java
@@ -3,6 +3,7 @@ package io.bootique.jdbc.instrumented.tomcat;
 import com.codahale.metrics.Gauge;
 import com.codahale.metrics.MetricRegistry;
 import io.bootique.jdbc.DataSourceListener;
+import io.bootique.metrics.MetricNaming;
 import org.apache.tomcat.jdbc.pool.ConnectionPool;
 
 import javax.sql.DataSource;
@@ -37,9 +38,11 @@ public class TomcatMetricsInitializer implements DataSourceListener {
         org.apache.tomcat.jdbc.pool.DataSource tomcat = (org.apache.tomcat.jdbc.pool.DataSource) dataSource;
         ConnectionPool pool = tomcat.getPool();
 
-        metricRegistry.register(MetricRegistry.name(getClass(), name, "active"), (Gauge<Integer>) () -> pool.getActive());
-        metricRegistry.register(MetricRegistry.name(getClass(), name, "idle"), (Gauge<Integer>) () -> pool.getIdle());
-        metricRegistry.register(MetricRegistry.name(getClass(), name, "waiting"), (Gauge<Integer>) () -> pool.getWaitCount());
-        metricRegistry.register(MetricRegistry.name(getClass(), name, "size"), (Gauge<Integer>) () -> pool.getSize());
+        MetricNaming naming = MetricNaming.forModule(JdbcTomcatInstrumentedModule.class);
+
+        metricRegistry.register(naming.name("Pool", name, "ActiveConnections"), (Gauge<Integer>) () -> pool.getActive());
+        metricRegistry.register(naming.name("Pool", name, "IdleConnections"), (Gauge<Integer>) () -> pool.getIdle());
+        metricRegistry.register(naming.name("Pool", name, "PendingConnections"), (Gauge<Integer>) () -> pool.getWaitCount());
+        metricRegistry.register(naming.name("Pool", name, "Size"), (Gauge<Integer>) () -> pool.getSize());
     }
 }