+++ bootique-jdbc-tomcat/src/test/java/io/bootique/jdbc/tomcat/JdbcTomcatModuleIT.java
@@ -14,7 +14,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
-public class TomcatJdbcModuleIT {
+public class JdbcTomcatModuleIT {
 
     @Rule
     public final BQTestFactory testFactory = new BQTestFactory();